+++ Slider/Assets/Prefabs/UI/Artifact/ArtifactInventoryItem.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b1443db5380833d48b886452d308d54c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
