--- Slider/Assets/Scripts/UI/Artifact/ArtifactScreenAnimator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1a129bd5e2e6c624d924132933083f3f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
