+++ Slider/Assets/Scripts/UI/Artifact/Screens.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0b81c2ebaf7b1454292b08dda3b2d1bd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
