+++ Slider/Assets/Scripts/UI/Artifact/Screens/ArtifactInventoryCollectible.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 961a8bf1ceae3414f8a89996e62b4f0a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
