+++ Slider/Assets/Scripts/UI/Artifact/Screens/UIArtifactInventory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 255c01956d30ee44fa0b09b736a31c8a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
