+++ Slider/Assets/Scripts/UI/Artifact/UIArtifact.cs
@@ -25,6 +25,11 @@ public class UIArtifact : MonoBehaviour
     
     public void Awake()
     {
+        Init();
+    }
+
+    public void Init()
+    {
         _instance = this;
     }
 
@@ -462,9 +467,9 @@ public class UIArtifact : MonoBehaviour
         return null;
     }
 
-    public ArtifactTileButton GetButton(int islandId){
+    public ArtifactTileButton GetButton(int islandId) { // this causes issues with UITracker and setting prefab parents for some reason...
 
-        foreach (ArtifactTileButton b in _instance.buttons)
+        foreach (ArtifactTileButton b in buttons)
         {
             if (b.islandId == islandId)
             {
@@ -516,7 +521,7 @@ public class UIArtifact : MonoBehaviour
 
     public void FlickerNewTiles()
     {
-        foreach (ArtifactTileButton b in _instance.buttons)
+        foreach (ArtifactTileButton b in buttons)
         {
             if (b.shouldFlicker)
             {
@@ -527,7 +532,7 @@ public class UIArtifact : MonoBehaviour
 
     public void FlickerAllOnce()
     {
-        foreach (ArtifactTileButton b in _instance.buttons)
+        foreach (ArtifactTileButton b in buttons)
         {
             b.Flicker(1);
         }
