+++ Slider/Assets/Scripts/UI/UIManager.cs
@@ -70,30 +70,27 @@ public class UIManager : MonoBehaviour
     {
         if (isGamePaused && pausePanel.activeSelf)
         {
-            Debug.Log("a");
             ResumeGame();
         }
-        else
+        else if (optionsPanel.activeSelf)
+        {
+            OpenPause();
+        }
+        else if (controlsPanel.activeSelf || advOptionsPanel.activeSelf)
         {
             // if in a pause sub-menu
-            if (controlsPanel.activeSelf || advOptionsPanel.activeSelf)
-            {
-                OpenOptions();
-            }
-            else
-            {
-                // if another menu is open (e.g. ocean shop)
-                if (IsUIOpen())
-                {
-                    // do nothing
-                    Debug.Log("Another menu is open, doing nothing..");
-                }
-                else 
-                {
-                    PauseGame();
-                    OpenPause();
-                }
-            }
+            OpenOptions();
+        }
+        else if (IsUIOpen())
+        {
+            // if another menu is open (e.g. ocean shop)
+            // do nothing
+            Debug.Log("Another menu is open, doing nothing..");
+        }
+        else 
+        {
+            PauseGame();
+            OpenPause();
         }
     }
