--- Slider/Assets/Sprites/Items/Items.ase.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b7698dcb6cf1d7146baf24a74efc3be6
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
