+++ Slider/Assets/Sprites/UI/artifact/artifact_world_map.ase.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3f119d636501b0f498f991d77db8daae
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
